# BookManagement/app/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Library API
  version: 1.0.0
  description: |
    Library Management API secured by Bearer JWT. Tokens are issued by an external auth server
    (issuer `http://localhost:3000`) and validated by this API. Endpoints indicate the required
    scopes or roles in their descriptions.
servers:
  - url: /
    description: Flask API Base URL (port 5000)
security:
  - SecurityToken: []
paths:
...
  /api/v1/users/:
    post:
      tags: [Users (v1)]
      summary: Create user
      description: Requires scope `users:write`.
      security:
        - SecurityToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
              required: [username]
      responses:
        '201':
          description: User created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [Users (v1)]
      summary: List users
      description: Requires scope `users:read`.
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{user_id}:
    get:
      tags: [Users (v1)]
      summary: Get user by ID
      description: Requires scope `users:read`.
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/users/{user_id}/borrows:
    get:
      tags: [Users (v1)]
      summary: List borrows for a user
      description: Requires scope `users:read`.
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated borrows for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/users/:
    get:
      tags:
        - "Users (v2)"
      summary: List users (v2)
      description: 'Requires scope `users:read`. Adds `extra_info: v2`.'  
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated users v2
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponseV2'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v2/users/{user_id}:
    get:
      tags:
        - "Users (v2)"
      summary: Get user by ID (v2)
      description: Requires scope `users:read`.
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/users/{user_id}/borrows:
    get:
      tags:
        - "Users (v2)"
      summary: List borrows for a user (v2)
      description: Requires scope `users:read`.
      security:
        - SecurityToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated borrows for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    SecurityToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use HTTP Bearer JWT issued by the external auth server (issuer `http://localhost:3000`).
        Required scopes and roles are indicated in each endpoint's description.
  parameters:
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      in: query
      name: per_page
      schema: { type: integer, minimum: 1, default: 10 }
...
    UserId:
      in: path
      name: user_id
      required: true
      schema: { type: integer }
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: missing_or_invalid_token }
              details: { type: string, example: JWT verification failed }
    Forbidden:
      description: Token lacks required scope or role.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: insufficient_scope }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    BorrowListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              user_id: { type: integer }
              copy_id: { type: integer }
              start_date: { type: string, format: date, nullable: true }
              end_date: { type: string, format: date, nullable: true }
        pagination:
          $ref: '#/components/schemas/Pagination'
        _links:
          type: object
          additionalProperties: true
    UserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id: { type: integer }
            username: { type: string }
            type: { type: string, enum: [user] }
        _links:
          type: object
          properties:
            self: { type: string }
            borrows: { type: string }
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              username: { type: string }
        pagination:
          $ref: '#/components/schemas/Pagination'
        _links:
          type: object
          additionalProperties: true
    UserListResponseV2:
      allOf:
        - $ref: '#/components/schemas/UserListResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  username: { type: string }
                  extra_info: { type: string, enum: [v2] }